% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{coef<-.pdKronecker}
\alias{coef<-.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
coef.pdKronecker(object, ...) <- value
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments.}

\item{value}{Numeric vector of coefficients.}
}
\value{
The modified `pdKronecker` object.
}
\description{
Assign Coefficients to pdKronecker
}
\details{
Assigns new coefficients to a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
