% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{corMatrix.pdKronecker}
\alias{corMatrix.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
corMatrix.pdKronecker(object, ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments.}
}
\value{
A correlation matrix with standard deviations as an attribute.
}
\description{
Extract Correlation Matrix from pdKronecker
}
\details{
Returns the correlation matrix of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
