% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerConstruct.R
\name{KroneckAux}
\alias{KroneckAux}
\title{File: pdKroneckerConstruct.txt}
\usage{
KroneckAux(pdx)
}
\arguments{
\item{pdx}{A list or object of class `pdKronecker` with initialized components.}
}
\value{
The rescaled `pdKronecker` object.
}
\description{
Rescale Coefficients for pdKronecker Components
}
\details{
Rescales coefficients of component `pdMat` objects in a `pdKronecker` object, ensuring
the first component (of class `pdIdent`) holds the sum of coefficients.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
\keyword{internal}
