% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{logDet.pdKronecker}
\alias{logDet.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
logDet.pdKronecker(object, ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments.}
}
\value{
Numeric; the log-determinant.
}
\description{
Compute Log-Determinant of pdKronecker
}
\details{
Computes the log-determinant of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
