% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{matrix<-.pdKronecker}
\alias{matrix<-.pdKronecker}
\title{Assign Matrix to pdKronecker}
\usage{
\method{matrix}{pdKronecker}(object) <- value

\method{matrix}{pdKronecker}(object) <- value
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{value}{A matrix to assign.}
}
\value{
The modified `pdKronecker` object.

The modified `pdKronecker` object.
}
\description{
Assigns a matrix to a `pdKronecker` object, updating its component matrices.

Assigns a matrix to a `pdKronecker` object, updating its component matrices.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
