% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.reStructU.R
\name{model.matrix.reStruct.U}
\alias{model.matrix.reStruct.U}
\title{File: model.matrix.reStructU.txt}
\usage{
\method{model.matrix}{reStruct.U}(object, data, contrast = NULL, ...)
}
\arguments{
\item{object}{An object inheriting from class `reStruct`, representing a random effects
structure as a list of `pdMat` objects.}

\item{data}{A data frame in which to evaluate the variables defined in `formula(object)`.}

\item{contrast}{An optional named list specifying contrasts for factor variables in `data`.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix obtained by binding together, column-wise, the model matrices for each
  element of `formula(object)`.
}
\description{
Create Model Matrix for reStruct Object
}
\details{
Replaces `model.matrix.reStruct` in the `nlme` package, creating a model matrix by
binding column-wise the model matrices for each element of a random effects structure.
}
\examples{
\dontrun{
  Pwr(fm1)
}
}
\seealso{
\code{\link[stats]{model.matrix}}
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
