% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{Names<-.pdKronecker}
\alias{Names<-.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
Names.pdKronecker(object, ...) <- value
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{value}{A character vector or list of names.}
}
\value{
The modified `pdKronecker` object.
}
\description{
Assign Names to pdKronecker
}
\details{
Assigns new names to a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
