% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{Names.pdKronecker}
\alias{Names.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
Names.pdKronecker(object, asList = FALSE, ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{asList}{Logical; if `TRUE`, returns a list of names.}

\item{...}{Additional arguments.}
}
\value{
A character vector or list of names.
}
\description{
Extract Names from pdKronecker
}
\details{
Returns the names of a `pdKronecker` object, optionally as a list.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
