% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmeUpdK-package.R
\docType{package}
\name{nlmeUpdK-package}
\alias{nlmeUpdK-package}
\alias{nlmeUpdK}
\title{Auxiliary Functions for pdKronecker Class}
\description{
nlmeUpdK: Companion Package to nlmeU for pdKronecker Class
}
\details{
The nlmeUpdK package is a companion to the nlmeU package, providing auxiliary functions
for the pdKronecker class introduced in Galecki and Burzykowski (2013). It includes
functionality for constructing and manipulating positive-definite Kronecker product matrices
and related methods for random effects structures. The package is under development.

Upon loading, nlmeUpdK modifies the nlme package namespace by replacing the internal
function model.matrix.reStruct with model.matrix.reStruct.U to support enhanced handling
of random effects structures, particularly for pdKronecker objects. This modification is
session-specific and reversible upon unloading the package.
}
\references{
Galecki, A., & Burzykowski, T. (2013). *Linear Mixed-Effects Models Using R: A Step-by-Step Approach*.
Springer.
}
\seealso{
\code{\link[nlmeU]{nlmeU}}
}
\author{
Andrzej Galecki \email{agalecki@umich.edu}, Tomasz Burzykowski \email{tomasz.burzykowski@uhasselt.be}
}
\keyword{package}
