% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerConstruct.R
\name{pdConstruct.pdKronecker}
\alias{pdConstruct.pdKronecker}
\title{Construct pdKronecker Object}
\usage{
\method{pdConstruct}{pdKronecker}(
  object,
  value = numeric(0),
  form = formula(object, TRUE),
  nam = Names(object, TRUE),
  data = sys.frame(sys.parent()),
  pdClass = sapply(object, function(el) class(el)[1]),
  ...
)
}
\arguments{
\item{object}{A `pdKronecker` object to be constructed.}

\item{value}{A list, matrix, or numeric vector specifying the initial values.}

\item{form}{A list of formulas for the component matrices.}

\item{nam}{A list of names for the component matrices.}

\item{data}{A data frame for evaluating formulas.}

\item{pdClass}{A character vector specifying the classes of component `pdMat` objects.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A constructed `pdKronecker` object.
}
\description{
Constructs a `pdKronecker` object from provided values, formulas, and names, ensuring
compatibility of dimensions and proper initialization of component `pdMat` objects.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
