% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerConstruct.R
\name{pdKronecker}
\alias{pdKronecker}
\title{File: pdKroneckerConstruct.txt}
\usage{
pdKronecker(
  value = numeric(0),
  form = NULL,
  nam = NULL,
  data = sys.frame(sys.parent())
)
}
\arguments{
\item{value}{A list with elements that are `pdMat` objects. The first element must be of
class `pdIdent`. Defaults to `numeric(0)`.}

\item{form}{A list of formulas specifying the structure of the matrices. Defaults to `NULL`.}

\item{nam}{A list of names for the matrices. Defaults to `NULL`.}

\item{data}{A data frame used by the source function. Defaults to `sys.frame(sys.parent())`.}
}
\value{
A `pdKronecker` object, also inheriting from class `pdMat`, representing a
  positive-definite Kronecker product matrix.
}
\description{
Construct a Positive-Definite Kronecker Product Matrix
}
\details{
This function constructs a `pdKronecker` object, representing a positive-definite matrix
based on the Kronecker product of underlying positive-definite matrices, each represented
by individual `pdMat` objects. The first element is of class `pdIdent` and acts as a scale
multiplier.
}
\examples{
\dontrun{
  runScript()
}
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
