% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{pdMatrix.pdKronecker}
\alias{pdMatrix.pdKronecker}
\title{Extract Matrix from pdKronecker}
\usage{
\method{pdMatrix}{pdKronecker}(object, factor = FALSE)

\method{pdMatrix}{pdKronecker}(object, factor = FALSE)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{factor}{Logical; if `TRUE`, returns the factor matrix.}
}
\value{
A matrix.

A matrix.
}
\description{
Returns the matrix representation of a `pdKronecker` object, optionally as a factor.

Returns the matrix representation of a `pdKronecker` object, optionally as a factor.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
