% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{solve.pdKronecker}
\alias{solve.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
\method{solve}{pdKronecker}(a, b, ...)
}
\arguments{
\item{a}{A `pdKronecker` object.}

\item{b}{Optional; not used.}

\item{...}{Additional arguments.}
}
\value{
A `pdKronecker` object representing the inverse.
}
\description{
Solve pdKronecker Object
}
\details{
Computes the inverse of a `pdKronecker` object by solving its component matrices.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
