% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{unlistFun}
\alias{unlistFun}
\title{File: pdKroneckerMethods.txt}
\usage{
unlistFun(namesList)
}
\arguments{
\item{namesList}{A list of character vectors containing names.}
}
\value{
A character vector of combined names.
}
\description{
Unlist Names for pdKronecker
}
\details{
Creates a character vector of names by combining names from a list using a Kronecker product-like structure.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
