% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{VarCorr.pdKronecker}
\alias{VarCorr.pdKronecker}
\title{Variance-Covariance Matrix for pdKronecker}
\usage{
\method{VarCorr}{pdKronecker}(x, sigma = 1, ...)
}
\arguments{
\item{x}{A `pdKronecker` object.}

\item{sigma}{Numeric; standard deviation scaling factor (default: 1).}

\item{...}{Additional arguments; `rdig` can be specified as an integer for the number of digits for rounding (default: 3).}
}
\value{
A matrix of variance-covariance components.
}
\description{
Returns the variance-covariance matrix components of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
