% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{coef<-.pdKronecker}
\alias{coef<-.pdKronecker}
\title{Assign Coefficients to pdKronecker}
\usage{
\method{coef}{pdKronecker}(object, ...) <- value

\method{coef}{pdKronecker}(object, ...) <- value
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments (currently ignored).}

\item{value}{Numeric vector of coefficients.}
}
\value{
The modified `pdKronecker` object.

The modified `pdKronecker` object.
}
\description{
Assigns new coefficients to a `pdKronecker` object.

Assigns new coefficients to a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
