% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{corMatrix.pdKronecker}
\alias{corMatrix.pdKronecker}
\title{Extract Correlation Matrix from pdKronecker}
\usage{
\method{corMatrix}{pdKronecker}(object, ...)

\method{corMatrix}{pdKronecker}(object, ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A correlation matrix with standard deviations as an attribute.

A correlation matrix with standard deviations as an attribute.
}
\description{
Returns the correlation matrix of a `pdKronecker` object.

Returns the correlation matrix of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
