% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{isInitialized.pdKronecker}
\alias{isInitialized.pdKronecker}
\title{Check Initialization of pdKronecker}
\usage{
\method{isInitialized}{pdKronecker}(object)

\method{isInitialized}{pdKronecker}(object)
}
\arguments{
\item{object}{A `pdKronecker` object.}
}
\value{
Logical; `TRUE` if all components are initialized.

Logical; `TRUE` if all components are initialized.
}
\description{
Checks if a `pdKronecker` object is initialized.

Checks if a `pdKronecker` object is initialized.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
