% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{isInitialized.pdKronecker}
\alias{isInitialized.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
isInitialized.pdKronecker(object)
}
\arguments{
\item{object}{A `pdKronecker` object.}
}
\value{
Logical; `TRUE` if all components are initialized.
}
\description{
Check Initialization of pdKronecker
}
\details{
Checks if a `pdKronecker` object is initialized.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
