% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{logDet.pdKronecker}
\alias{logDet.pdKronecker}
\title{Compute Log-Determinant of pdKronecker}
\usage{
\method{logDet}{pdKronecker}(object, ...)

\method{logDet}{pdKronecker}(object, ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Numeric; the log-determinant.

Numeric; the log-determinant.
}
\description{
Computes the log-determinant of a `pdKronecker` object.

Computes the log-determinant of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
