% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{Names<-.pdKronecker}
\alias{Names<-.pdKronecker}
\title{Assign Names to pdKronecker}
\usage{
\method{Names}{pdKronecker}(object, ...) <- value

\method{Names}{pdKronecker}(object, ...) <- value
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{...}{Additional arguments (currently ignored).}

\item{value}{A character vector or list of names.}
}
\value{
The modified `pdKronecker` object.

The modified `pdKronecker` object.
}
\description{
Assigns new names to a `pdKronecker` object.

Assigns new names to a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
