% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{pdFactor.pdKronecker}
\alias{pdFactor.pdKronecker}
\title{Extract Factor of pdKronecker}
\usage{
\method{pdFactor}{pdKronecker}(object)

\method{pdFactor}{pdKronecker}(object)
}
\arguments{
\item{object}{A `pdKronecker` object.}
}
\value{
A matrix representing the factor.

A matrix representing the factor.
}
\description{
Returns the factor of a `pdKronecker` matrix.

Returns the factor of a `pdKronecker` matrix.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
