% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{pdMatrix.pdKronecker}
\alias{pdMatrix.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
pdMatrix.pdKronecker(object, factor = FALSE)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{factor}{Logical; if `TRUE`, returns the factor matrix.}
}
\value{
A matrix.
}
\description{
Extract Matrix from pdKronecker
}
\details{
Returns the matrix representation of a `pdKronecker` object, optionally as a factor.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
