% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{solve.pdKronecker}
\alias{solve.pdKronecker}
\title{Solve pdKronecker Object}
\usage{
\method{solve}{pdKronecker}(a, b, ...)
}
\arguments{
\item{a}{A `pdKronecker` object.}

\item{b}{Optional; not used.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A `pdKronecker` object representing the inverse.
}
\description{
Computes the inverse of a `pdKronecker` object by solving its component matrices.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
