% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{[.pdKronecker}
\alias{[.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
\method{[}{pdKronecker}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{A `pdKronecker` object.}

\item{i}{Row indices.}

\item{j}{Column indices.}

\item{drop}{Logical; if `TRUE`, drops dimensions when possible.}

\item{...}{Additional arguments.}
}
\value{
A matrix or a `pdKronecker` object.
}
\description{
Subset pdKronecker Object
}
\details{
Subsets a `pdKronecker` object, returning either a matrix or a new `pdKronecker` object
based on the indices provided.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
