% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{summary.pdKronecker}
\alias{summary.pdKronecker}
\title{Summarize pdKronecker Object}
\usage{
\method{summary}{pdKronecker}(object, structName = "pdKronecker", ...)
}
\arguments{
\item{object}{A `pdKronecker` object.}

\item{structName}{Character; name of the structure (default: "pdKronecker").}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A summary object of class `summary.pdMatX`.
}
\description{
Summarize pdKronecker Object
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
