% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdKroneckerMethods.R
\name{VarCorr.pdKronecker}
\alias{VarCorr.pdKronecker}
\title{File: pdKroneckerMethods.txt}
\usage{
VarCorr.pdKronecker(x, sigma = 1, rdig = 3)
}
\arguments{
\item{x}{A `pdKronecker` object.}

\item{sigma}{Numeric; standard deviation scaling factor (default: 1).}

\item{rdig}{Integer; number of digits for rounding (default: 3).}
}
\value{
A matrix of variance-covariance components.
}
\description{
Variance-Covariance Matrix for pdKronecker
}
\details{
Returns the variance-covariance matrix components of a `pdKronecker` object.
}
\author{
Andrzej Galecki, Tomasz Burzykowski
}
